# api/openapi/aor.v1.yaml
openapi: 3.2.0
info:
  title: Agent Organization Runtime API
  version: 1.0.0
  description: Deterministic, tenant-scoped control API for SPEC 2.0.0.
servers:
  - url: https://aor.example.invalid
security:
  - oidc: [aor]
  - mutualTLS: []
...
paths:
  /v1/model-providers:
    get:
      tags: [Settings]
      operationId: listModelProviders
      responses:
        '200':
          description: Configured provider metadata without endpoints or credential references
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ModelProviderPage'}
        default: {$ref: '#/components/responses/Problem'}
...
  /v1/settings/model-routes:
    get:
      tags: [Settings]
      operationId: getModelRouteSettings
      responses:
        '200': {$ref: '#/components/responses/ModelRouteSettings'}
        default: {$ref: '#/components/responses/Problem'}
    put:
      tags: [Settings]
      operationId: putModelRouteSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [modelRoutes]
              properties:
                modelRoutes: {$ref: '#/components/schemas/ModelRoutes'}
      responses:
        '200': {$ref: '#/components/responses/ModelRouteSettings'}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example.invalid/.well-known/openid-configuration
    mutualTLS:
      type: mutualTLS
  headers:
    ETag:
      description: Opaque aggregate version tag.
      schema: {type: string, pattern: '^"[A-Za-z0-9._:-]+"$'}
...
  schemas:
    ProjectModelRoute:
      type: object
      additionalProperties: false
      required: [provider, model, reasoningEffort, maxOutputTokens, temperature, providerPolicy, cachePolicy, worstCaseCostMicros, maxAttempts]
      properties:
        provider: {type: string, minLength: 1, maxLength: 128}
        model: {type: string, minLength: 1, maxLength: 256}
        reasoningEffort: {enum: ['', none, low, medium, high, xhigh, max]}
        maxOutputTokens: {type: integer, minimum: 1, maximum: 1000000}
        temperature: {type: number, minimum: 0, maximum: 2}
        seed: {type: integer, format: int64}
        providerPolicy: {type: string, minLength: 1, maxLength: 256}
        cachePolicy: {type: string, minLength: 1, maxLength: 128}
        worstCaseCostMicros: {type: integer, format: int64, minimum: 0}
        maxAttempts: {type: integer, minimum: 1, maximum: 5}
    ModelRoutes:
      type: object
      additionalProperties: false
      required: [GOAL_PROPOSER, GOAL_CHALLENGER, PLAN_SUPERVISOR, MODULE_PLANNER, EXECUTOR, MODULE_AUDITOR, GLOBAL_AUDITOR, KNOWLEDGE_CURATOR]
      properties:
        GOAL_PROPOSER: {$ref: '#/components/schemas/ProjectModelRoute'}
        GOAL_CHALLENGER: {$ref: '#/components/schemas/ProjectModelRoute'}
        PLAN_SUPERVISOR: {$ref: '#/components/schemas/ProjectModelRoute'}
        MODULE_PLANNER: {$ref: '#/components/schemas/ProjectModelRoute'}
        EXECUTOR: {$ref: '#/components/schemas/ProjectModelRoute'}
        MODULE_AUDITOR: {$ref: '#/components/schemas/ProjectModelRoute'}
        GLOBAL_AUDITOR: {$ref: '#/components/schemas/ProjectModelRoute'}
        KNOWLEDGE_CURATOR: {$ref: '#/components/schemas/ProjectModelRoute'}
    ModelRouteSettings:
      type: object
      additionalProperties: false
      required: [modelRoutes, version]
      properties:
        modelRoutes: {$ref: '#/components/schemas/ModelRoutes'}
        version: {type: integer, format: int64, minimum: 0}
...
    ModelProvider:
      type: object
      additionalProperties: false
      required: [id, provider, models, inputMicrosPerToken, outputMicrosPerToken, supportsStreaming, supportsToolCalls, supportsJsonSchema, supportsSeed, supportsPromptCaching, maxInputTokens, maxOutputTokens, allowedDataClassifications, dataResidency, retentionPolicy, modalities]
      properties:
        id: {type: string, minLength: 1, maxLength: 128}
        provider: {type: string, minLength: 1, maxLength: 128}
        models:
          type: array
          minItems: 1
          uniqueItems: true
          items: {type: string, minLength: 1, maxLength: 256}
        inputMicrosPerToken: {type: integer, format: int64, minimum: 0}
        outputMicrosPerToken: {type: integer, format: int64, minimum: 0}
        supportsStreaming: {type: boolean}
        supportsToolCalls: {type: boolean}
        supportsJsonSchema: {type: boolean}
        supportsSeed: {type: boolean}
        supportsPromptCaching: {type: boolean}
        maxInputTokens: {type: integer, minimum: 1}
        maxOutputTokens: {type: integer, minimum: 1}
        allowedDataClassifications:
          type: array
          minItems: 1
          uniqueItems: true
          items: {type: string, minLength: 1, maxLength: 64}
        dataResidency:
          type: array
          minItems: 1
          uniqueItems: true
          items: {type: string, minLength: 1, maxLength: 128}
        retentionPolicy: {type: string, minLength: 1, maxLength: 256}
        modalities:
          type: array
          minItems: 1
          uniqueItems: true
          items: {type: string, minLength: 1, maxLength: 128}
    ModelProviderPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/ModelProvider'}
...
  responses:
    ModelRouteSettings:
      description: Effective tenant model routes and persisted settings version
      headers:
        ETag: {$ref: '#/components/headers/ETag'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ModelRouteSettings'}
...
    Problem:
      description: Redacted AOR Problem Details
      content:
        application/problem+json:
          schema: {$ref: '../json-schema/aor-error.v1.schema.json'}
...